remove master search
